TI-76 - refacto(prospect classes): confirm the suggestions for ProspectManager

+++ smartimmo/src/main/java/com/gangdestrois/smartimmo/domain/prospect/ProspectStatistic.java
@@ -1,4 +1,4 @@
 package com.gangdestrois.smartimmo.domain.prospect;
 
-public record ProspectStatistic (String dataProspect, Double count){
+public record ProspectStatistic (String dataProspect, long count){
 }
